+++ src-tauri/src/lib_commands/app_api/mod.rs
@@ -1,0 +1,13 @@
+use super::*;
+
+mod core;
+mod navidrome;
+mod remote;
+mod integration;
+mod analysis;
+
+pub(crate) use core::*;
+pub(crate) use navidrome::*;
+pub(crate) use remote::*;
+pub(crate) use integration::*;
+pub(crate) use analysis::*;
