+++ src-tauri/src/lib_commands/cache/mod.rs
@@ -1,0 +1,9 @@
+use super::*;
+
+mod offline;
+mod downloads;
+mod hot;
+
+pub(crate) use offline::*;
+pub(crate) use downloads::*;
+pub(crate) use hot::*;
