+++ src-tauri/src/lib_commands/ui/mod.rs
@@ -1,0 +1,7 @@
+use super::*;
+
+mod mini;
+mod bandsintown;
+
+pub(crate) use mini::*;
+pub(crate) use bandsintown::*;
